+++ pkg/ifuzz/ifuzz_test.go
@@ -36,7 +36,7 @@ func TestMode(t *testing.T) {
 }
 
 func TestDecode(t *testing.T) {
-	seed := int64(time.Now().UnixNano())
+	seed := time.Now().UnixNano()
 	if os.Getenv("TRAVIS") != "" {
 		seed = 0 // required for deterministic coverage reports
 	}
